ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
deepface==0.0.93
numpy==1.26.4
faiss-cpu==1.8.0
Pillow==10.3.0
httpx==0.27.0
aiofiles==23.2.1
motor==3.4.0
pymongo==4.7.2
python-dotenv==1.0.1
opencv-python-headless==4.9.0.80
tf-keras==2.16.0
